requirements.txt
fastapi==0.128.0
PyYAML==6.0.3
uvicorn==0.40.0
uvloop==0.22.1

